requirements.txt
streamlit
openai
langchain
faiss-cpu
pickle-mixin
PyMuPDF